Factored out player set into separate file, moved player into sim module
