+++ src/main.rs
@@ -41,7 +41,6 @@ mod config;
 mod core;
 mod input;
 mod network;
-mod player;
 mod render;
 mod server;
 mod simple;
